정수현/Dust-Health-AI_v11/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
colorama==0.4.6
psycopg2
django-cors-headers
mysqlclient
requests
pandas